[refactor] Screen 컴포넌트 접미사 추가
#92
