+++ src/screens/AddTrainingScreen/styled.ts
